+++ src/pages/Board.jsx
@@ -1,0 +1,4 @@
+export const BoardPage = () => {
+  return <></>;
+};
+export default BoardPage;
