--- src/pages/ScreenPage.jsx
@@ -1,4 +1,0 @@
-export const ScreenPage = () => {
-  return <></>;
-};
-export default ScreenPage;
