+++ src/routes/PublicPage.jsx
@@ -1,7 +1,7 @@
 import { Navigate } from 'react-router-dom';
 
-const PublicPage = ({ component: Component, redirectTo = '/' }) => {
+const PublicPage = ({ component: Component, redirectTo = '/home' }) => {
   const isLoggedIn = false;
-  return isLoggedIn ? <Navigate to={redirectTo} /> : Component;
+  return isLoggedIn ? <Navigate to={redirectTo} /> : <Component />;
 };
 export default PublicPage;
